修改用户权限

+++ biz/wx-dss/src/main/java/com/xiaosi/wx/mapper/SysRoleMapper.java
@@ -18,5 +18,5 @@ public interface SysRoleMapper {
 
     int updateByPrimaryKey(SysRole record);
 
-    List<SysRole> selectByUserId(Integer userId);
+    List<SysRole> selectByUserId(Long userId);
 }
